Merge pull request #5 from ArtemYo8283/feature/task5

+++ CoffeeMachine/app/src/main/java/com/nix/summer/finall/core/interactors/TakeMoneyInteractor.kt
@@ -6,5 +6,5 @@ import com.nix.summer.finall.core.entities.Status
 import com.nix.summer.finall.data.repositories.ActionRepository
 
 class TakeMoneyInteractor(private val repository: ActionRepository) {
-    operator fun invoke(): Int = repository.take()
+    operator fun invoke(): Double = repository.take()
 }
